+++ src/biz/orgin/minecraft/hothgenerator/HothUtils.java
@@ -179,47 +179,46 @@ public class HothUtils
 		}
 	}
 	
-	public static void setPos(byte[][] chunk, int x, int y, int z, Material material)
+	public static void setPos(short[][] chunk, int x, int y, int z, Material material)
 	{
 		int type = MaterialManager.toID(material);
-		HothUtils.setPos(chunk, x, y, z, type);
+		HothUtils.setPos(chunk, x, y, z, (short)type);
 	}
 	
-	public static void setPos(byte[][] chunk, int x, int y, int z, int type)
+	public static void setPos(short[][] chunk, int x, int y, int z, int type)
 	{
-		
 		int sub = y/16;
 		int rely = y-(sub*16);
 		
 		if(chunk[sub]==null)
 		{
-			chunk[sub] = new byte[16*16*16];
+			chunk[sub] = new short[16*16*16];
 		}
 		
-		HothUtils.setBlock(chunk[sub], x,rely,z, (byte)type);
+		HothUtils.setBlock(chunk[sub], x,rely,z, (short)type);
 
 	}
 	
-	public static void setBlock(byte[] subchunk, int x, int y, int z, byte blkid)
+	public static void setBlock(short[] subchunk, int x, int y, int z, short blkid)
 	{
 		subchunk[((y) << 8) | (z << 4) | x] = blkid;
 	}
 
-	public static byte getPos(byte[][] chunk, int x, int y, int z)
+	public static short getPos(short[][] chunk, int x, int y, int z)
 	{
 		int sub = y/16;
 		int rely = y-(sub*16);
 		
 		if(chunk[sub]==null)
 		{
-			chunk[sub] = new byte[16*16*16];
+			chunk[sub] = new short[16*16*16];
 		}
 		
 		return HothUtils.getBlock(chunk[sub], x,rely,z);
 
 	}
 	
-	public static void replaceTop(byte[][] chunk, byte from1, byte from2, byte to, int maxy)
+	public static void replaceTop(short[][] chunk, byte from1, byte from2, byte to, int maxy)
 	{
 		for(int x=0;x<16;x++)
 		{
@@ -228,7 +227,7 @@ public class HothUtils
 				int y = HothUtils.getMaxY(chunk, x, z, maxy);
 				if(y>0)
 				{
-					byte old = HothUtils.getPos(chunk, x, y, z);
+					short old = HothUtils.getPos(chunk, x, y, z);
 					if(old==from1 || old==from2)
 					{
 						HothUtils.setPos(chunk, x, y, z, to);
@@ -238,7 +237,7 @@ public class HothUtils
 		}
 	}
 	
-	private static int getMaxY(byte[][] chunk, int x, int z, int maxy)
+	private static int getMaxY(short[][] chunk, int x, int z, int maxy)
 	{
 		for(int i=(maxy-1);i>0;i--)
 		{
@@ -252,7 +251,7 @@ public class HothUtils
 		return 0;
 	}
 	
-	private static int getRawPos(byte[][] chunk, int x, int y, int z)
+	private static int getRawPos(short[][] chunk, int x, int y, int z)
 	{
 		int sub = y/16;
 		int rely = y-(sub*16);
@@ -265,7 +264,7 @@ public class HothUtils
 		return HothUtils.getBlock(chunk[sub], x,rely,z);
 	}
 
-	public static byte getBlock(byte[] subchunk, int x, int y, int z)
+	public static short getBlock(short[] subchunk, int x, int y, int z)
 	{
 		return subchunk[((y) << 8) | (z << 4) | x];
 	}
